requirements.txt
# Dependencies for AI4Physics Assignment 2
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

matplotlib>=3.8.0
seaborn>=0.13.0

plotly>=5.17.0
statsmodels>=0.14.0
numba>=0.58.0

scikit-learn>=1.3.0
xgboost>=3.0.0
tensorflow-cpu>=2.19.0

tqdm>=4.66.0
click>=8.1.0

jupyter>=1.0.0
ipykernel>=6.25.0



